+++ packages/node-fetch/src/Blob.ts
@@ -1,6 +1,6 @@
 /* eslint-disable @typescript-eslint/no-unsafe-declaration-merging */
 import { PonyfillReadableStream } from './ReadableStream.js';
-import { uint8ArrayToArrayBuffer } from './utils.js';
+import { isArrayBufferView } from './utils.js';
 
 interface BlobOptions {
   /**
@@ -20,9 +20,7 @@ function getBlobPartAsBuffer(blobPart: Exclude<BlobPart, Blob>) {
     return Buffer.from(blobPart);
   } else if (Buffer.isBuffer(blobPart)) {
     return blobPart;
-  } else if (blobPart instanceof Uint8Array) {
-    return Buffer.from(blobPart);
-  } else if ('buffer' in blobPart) {
+  } else if (isArrayBufferView(blobPart)) {
     return Buffer.from(blobPart.buffer, blobPart.byteOffset, blobPart.byteLength);
   } else {
     return Buffer.from(blobPart);
@@ -30,7 +28,7 @@ function getBlobPartAsBuffer(blobPart: Exclude<BlobPart, Blob>) {
 }
 
 function isBlob(obj: any): obj is Blob {
-  return obj != null && typeof obj === 'object' && obj.arrayBuffer != null;
+  return obj != null && obj.arrayBuffer != null;
 }
 
 // Will be removed after v14 reaches EOL
@@ -62,7 +60,7 @@ export class PonyfillBlob implements Blob {
   }
 
   arrayBuffer() {
-    return this.buffer().then(uint8ArrayToArrayBuffer);
+    return this.buffer().then(buf => buf.buffer);
   }
 
   async text() {
@@ -70,7 +68,7 @@ export class PonyfillBlob implements Blob {
     for (const blobPart of this.blobParts) {
       if (typeof blobPart === 'string') {
         text += blobPart;
-      } else if ('text' in blobPart) {
+      } else if (isBlob(blobPart)) {
         text += await blobPart.text();
       } else {
         const buf = getBlobPartAsBuffer(blobPart);
@@ -87,9 +85,7 @@ export class PonyfillBlob implements Blob {
         size += Buffer.byteLength(blobPart);
       } else if (isBlob(blobPart)) {
         size += blobPart.size;
-      } else if ('length' in blobPart) {
-        size += (blobPart as Buffer).length;
-      } else if ('byteLength' in blobPart) {
+      } else if (isArrayBufferView(blobPart)) {
         size += blobPart.byteLength;
       }
     }